+++ CHANGELOG.md
@@ -4,13 +4,18 @@ All notable changes to the "log-highlighter" extension will be documented in thi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.0] 2025-01-28
+
+- Added dynamic search
+- Added hideNonMatchingLines property for hide no matching lines
+
 ## [1.0.2] 2025-01-25
 
 - Fix regexp bug
 
 ## [1.0.1] 2025-01-25
 
-- Filter by regexp using /~<regexp>~/
+- Filter by regexp using wrapped by `/~<regexp>~/`
 
 ## [1.0.0] 2025-01-25